backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
passlib==1.7.4
bcrypt==4.0.1